Change workaround for conflicting types between AK and macOS
This is super ugly, but tries to hide the ugliness in one location so
that not every user of macOS system headers has to do this workaround.
The `Duration` and `FixedPoint` types conflict with types of the same
name from macOS system headers. Originally, I was trying to define
AK_DONT_REPLACE_STD, which keeps the AK types in the AK namespace. But
once LibCore/LibWeb come into the picture, there are too many locations
within those libraries that assume AK types are in the global namespace
(including these 2 conflicting types). So the approach here is to add a
header for every system header we want to include (just Cocoa.h so far),
and then wrap the real include statement with the #define/#undef trick.