Draw the rendered page to the viewport with NSImage
An upcoming commit will mutate the coordinate system of the web view to
work better with NSScrollView. Draw onto the viewport with NSImage so
those mutations are handled - CGImage is lower-level and will not be
aware of those mutations.

Also add a comment explaning why CGImage was being used at all in the
first place.