Place the web view inside a scrolling container
The NSScrollView has two components, a document view and a content view.
The document view is effectively the "real" view, which applications are
meant to paint / add subviews into. The content view is the cropped view
of the document view that is actually visible.

So normally, we'd paint the entire web page onto the document view, and
the content view would scroll around that view. However, we don't paint
the entire web page - we specifically only paint the visible portion.

To accommodate this, we trick the NSScrollView by creating an entirely
empty document view, sized to the layout size provided to us by LibWeb.
This allows the scroll bars to appear with an appropriate size. We then
make the web view the content view, and paint the portion of the web
page that we have accordingly.

FIXME: Scrolling speed is pretty bad. The larger the scroll area, the
slower we scroll. If the scroll area is very small, we scroll incredibly
fast. Not a very good experience, so we should improve on this.