Add a helper class to convert a mouse NSEvent to a WebContent event
This is a bit unfortunate in that we include 2 headers that we should
never include; one from LibGUI and one from Kernel. But that happens to
be where the types needed for WebContent are stored.

Note that we have to explicitly map ctrl+left click to a right click. It
seems this should already be happening by AppKit, but I haven't seen it
happen in practice:
https://stackoverflow.com/questions/227797/can-you-detect-a-ctrl-click-context-menu-request-in-the-mousedown-event