Convert Objective-C selectors to camel case
Mixing snake case and camel case sticks out like a sore thumb in these
interfaces. We were using snake case for anything self-written, and
camel case for anything overridden / referenced from standard types.

However, the man reason is that there are some built-in selectors in
camel case that we will need to add self-written defintions of. For
example, selectAll: is a standard action selector, and when we write
the definition of a select-all context menu item, it *must* have the
camel cased name for the cmd+a shortcut to be automatically hooked up.