Ensure all navigations go through the TabController
The TabController is where the history navigation actions are handled
(forward, back, reload). Thus it will own the Browser::History object.
This ensures all navigations go through the TabController so it may
augment the history object accordingly.

It would be nice to *really* ensure this with AK::Badge (i.e. so only
TabController may call the web view's load: selector). But we currently
can't use AK::Badge with Objective-C interfaces.