Do not constantly send a new viewport rect while live-resizing
The window doesn't repaint anyways until the live resize ends, so we
would send a million IPC messages to WebContent with our new window size
and not be able to do anything about the updated layout. This works
around an issue where resizing a window constantly for ~30 seconds would
eventually make the browser hang).