Receive mouse move events when the web view is not the first responder
We currently had to force the web view to be the first responder for it
to receive mouse move events. This had unfortunate consequences such as
taking the first responder status away from the URL bar if your mouse
entered the web view while you were typing.

Instead, it turns out we can create the tracking area once, with options
enabled to always receive the mouse move events, even when the web view
is not the first responder. There is also an option to automatically
update the tracking area when the visible rect changes; so we do not
need to manually update the tracking area ourselves.