Implement alert, confirm, and prompt
There's a bit more involved here than just [NSAlert runModal], because
that would block the main thread completely. We need to be able to
interact with the dialog while it is running (mostly for WebDriver). If
the main thread were blocked, we would not be able to handle those
requests.